amount의 type을 전부 Long으로 수정

See merge request s09-fintech-finance-sub2/S09P22A108!120

Former-commit-id: c77ec8bcc941ee61303ebba0bc817e630ed774a5

+++ backend/src/main/java/com/sss/tally/api/calculate/dto/CalculateDto.java
@@ -141,9 +141,9 @@ public class CalculateDto {
 
 		private String paymentName;
 
-		private int myAmount;
+		private Long myAmount;
 
-		private int allAmount;
+		private Long allAmount;
 
 		private LocalDateTime paymentDate;
